more fixes for windows build
git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta/@54805 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 0453aa12c8e37b481aeb33325d66e304d7ec8ba6

+++ rosetta_source/src/protocols/loops/loop_closure/kinematic_closure/KinematicMover.cc
@@ -886,7 +886,7 @@ KinematicMover::set_defaults() {
 	void
 	KinematicMover::insert_sampled_torsion_string_into_taboo_map( std::string const & ts )
 	{
-		if (sequence_.size() > 0 && taboo_map_.size() > pow(4, sequence_.size()) * taboo_map_max_capacity_) {
+		if (sequence_.size() > 0 && taboo_map_.size() > pow(4.0, (double) sequence_.size()) * taboo_map_max_capacity_) {
 			TR << "Taboo map filled to " << taboo_map_max_capacity_*100 << "% capacity, clearing the map to avoid almost-endless loops. " << taboo_map_.size() << std::endl;
 			taboo_map_.clear();
 		}
